🔨 refactor: 로직 함수로 분리
